+++ packages/contracts/src/factories/VAnchorBase__factory.ts
@@ -12,15 +12,15 @@ const _abi = [
     inputs: [
       {
         indexed: false,
-        internalType: "uint256",
+        internalType: "uint64",
         name: "chainID",
-        type: "uint256",
+        type: "uint64",
       },
       {
         indexed: false,
-        internalType: "uint256",
+        internalType: "uint32",
         name: "latestLeafIndex",
-        type: "uint256",
+        type: "uint32",
       },
       {
         indexed: false,
@@ -37,15 +37,15 @@ const _abi = [
     inputs: [
       {
         indexed: false,
-        internalType: "uint256",
+        internalType: "uint64",
         name: "chainID",
-        type: "uint256",
+        type: "uint64",
       },
       {
         indexed: false,
-        internalType: "uint256",
+        internalType: "uint32",
         name: "latestLeafIndex",
-        type: "uint256",
+        type: "uint32",
       },
       {
         indexed: false,
@@ -268,9 +268,9 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint256",
+        internalType: "uint64",
         name: "",
-        type: "uint256",
+        type: "uint64",
       },
     ],
     name: "currentNeighborRootIndex",
@@ -300,9 +300,9 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint256",
+        internalType: "uint64",
         name: "",
-        type: "uint256",
+        type: "uint64",
       },
     ],
     name: "edgeExistsForChain",
@@ -319,17 +319,17 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint256",
+        internalType: "uint64",
         name: "",
-        type: "uint256",
+        type: "uint64",
       },
     ],
     name: "edgeIndex",
     outputs: [
       {
-        internalType: "uint256",
+        internalType: "uint32",
         name: "",
-        type: "uint256",
+        type: "uint32",
       },
     ],
     stateMutability: "view",
@@ -346,9 +346,9 @@ const _abi = [
     name: "edgeList",
     outputs: [
       {
-        internalType: "uint256",
+        internalType: "uint64",
         name: "chainID",
-        type: "uint256",
+        type: "uint64",
       },
       {
         internalType: "bytes32",
@@ -356,9 +356,9 @@ const _abi = [
         type: "bytes32",
       },
       {
-        internalType: "uint256",
+        internalType: "uint32",
         name: "latestLeafIndex",
-        type: "uint256",
+        type: "uint32",
       },
     ],
     stateMutability: "view",
@@ -416,9 +416,9 @@ const _abi = [
       {
         components: [
           {
-            internalType: "uint256",
+            internalType: "uint64",
             name: "chainID",
-            type: "uint256",
+            type: "uint64",
           },
           {
             internalType: "bytes32",
@@ -426,9 +426,9 @@ const _abi = [
             type: "bytes32",
           },
           {
-            internalType: "uint256",
+            internalType: "uint32",
             name: "latestLeafIndex",
-            type: "uint256",
+            type: "uint32",
           },
         ],
         internalType: "struct LinkableTree.Edge[]",
@@ -468,9 +468,9 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint256",
+        internalType: "uint64",
         name: "_chainID",
-        type: "uint256",
+        type: "uint64",
       },
     ],
     name: "hasEdge",
@@ -547,9 +547,9 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint256",
+        internalType: "uint64",
         name: "neighborChainID",
-        type: "uint256",
+        type: "uint64",
       },
       {
         internalType: "bytes32",
@@ -712,9 +712,9 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint256",
+        internalType: "uint64",
         name: "",
-        type: "uint256",
+        type: "uint64",
       },
       {
         internalType: "uint32",
@@ -877,9 +877,9 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint256",
+        internalType: "uint64",
         name: "sourceChainID",
-        type: "uint256",
+        type: "uint64",
       },
       {
         internalType: "bytes32",
@@ -887,9 +887,9 @@ const _abi = [
         type: "bytes32",
       },
       {
-        internalType: "uint256",
+        internalType: "uint32",
         name: "leafIndex",
-        type: "uint256",
+        type: "uint32",
       },
     ],
     name: "updateEdge",
